requirements (1).txt
streamlit
pandas
google-generativeai
pillow
duckduckgo-search